// GroceryStore/src/app/models/product.models.ts
export interface Product {
    productId:         number;
    name:              string;
    description:       string;
    category:          string;
    availableQuantity: number;
    imageUrl:          string;
    price:             number;
    discount:          number;
    specification:     string;
}